Update startTime and endTime types

+++ TripAnalysisUI/src/main/java/com/drivequant/drivekit/tripanalysis/workinghours/activity/WorkingHoursActivity.kt
@@ -157,7 +157,7 @@ class WorkingHoursActivity : AppCompatActivity() {
                         dataUpdated(true)
                     }
 
-                    override fun onHoursUpdated(start: Float, end: Float) {
+                    override fun onHoursUpdated(start: Double, end: Double) {
                         dataUpdated(true)
                     }
                 })
